+++ lib/core/decision_service/index.tests.js
@@ -32,7 +32,7 @@ import OptimizelyUserContext from '../../optimizely_user_context';
 import projectConfig, { createProjectConfig } from '../../project_config/project_config';
 import AudienceEvaluator from '../audience_evaluator';
 import errorHandler from '../../plugins/error_handler';
-import eventDispatcher from '../../event_processor/default_dispatcher.browser';
+import eventDispatcher from '../../event_processor/event_dispatcher/default_dispatcher.browser';
 import * as jsonSchemaValidator from '../../utils/json_schema_validator';
 import { getMockProjectConfigManager } from '../../tests/mock/mock_project_config_manager';
