+++ lib/event_processor/event_dispatcher/default_dispatcher.browser.spec.ts
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { vi, expect, it, describe, afterAll } from 'vitest';
 
 vi.mock('./default_dispatcher', () => {
@@ -20,30 +21,29 @@ vi.mock('./default_dispatcher', () => {
   return { DefaultEventDispatcher };
 });
 
-vi.mock('../utils/http_request_handler/node_request_handler', () => {
-  const NodeRequestHandler = vi.fn();
-  return { NodeRequestHandler };
+vi.mock('../../utils/http_request_handler/browser_request_handler', () => {
+  const BrowserRequestHandler = vi.fn();
+  return { BrowserRequestHandler };
 });
 
 import { DefaultEventDispatcher } from './default_dispatcher';
-import { NodeRequestHandler } from '../utils/http_request_handler/node_request_handler';
-import eventDispatcher from './default_dispatcher.node';
+import { BrowserRequestHandler } from '../../utils/http_request_handler/browser_request_handler';
+import eventDispatcher from './default_dispatcher.browser';
 
 describe('eventDispatcher', () => {
-  const MockNodeRequestHandler = vi.mocked(NodeRequestHandler);
-  const MockDefaultEventDispatcher = vi.mocked(DefaultEventDispatcher);
-
   afterAll(() => {
     MockDefaultEventDispatcher.mockReset();
-    MockNodeRequestHandler.mockReset();
-  })
+    MockBrowserRequestHandler.mockReset();
+  });
+  const MockBrowserRequestHandler = vi.mocked(BrowserRequestHandler);
+  const MockDefaultEventDispatcher = vi.mocked(DefaultEventDispatcher);
 
   it('creates and returns the instance by calling DefaultEventDispatcher', () => {
     expect(Object.is(eventDispatcher, MockDefaultEventDispatcher.mock.instances[0])).toBe(true);
   });
 
-  it('uses a NodeRequestHandler', () => {
+  it('uses a BrowserRequestHandler', () => {
     expect(Object.is(eventDispatcher, MockDefaultEventDispatcher.mock.instances[0])).toBe(true);
-    expect(Object.is(MockDefaultEventDispatcher.mock.calls[0][0], MockNodeRequestHandler.mock.instances[0])).toBe(true);
+    expect(Object.is(MockDefaultEventDispatcher.mock.calls[0][0], MockBrowserRequestHandler.mock.instances[0])).toBe(true);
   });
 });
