+++ lib/event_processor/event_dispatcher/event_dispatcher.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024 Optimizely
+ * Copyright 2022, 2024, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { EventDispatcher } from './event_dispatcher';
-import { NodeRequestHandler } from '../utils/http_request_handler/node_request_handler';
-import { DefaultEventDispatcher } from './default_dispatcher';
+import { EventBatch } from "../event_builder/log_event";
 
-const eventDispatcher: EventDispatcher = new DefaultEventDispatcher(new NodeRequestHandler());
+export type EventDispatcherResponse = {
+  statusCode?: number  
+}
 
-export default eventDispatcher;
+export interface EventDispatcher {
+  dispatchEvent(event: LogEvent): Promise<EventDispatcherResponse>
+}
+
+export interface LogEvent {
+  url: string
+  httpVerb: 'POST' | 'PUT' | 'GET' | 'PATCH'
+  params: EventBatch,
+}
