+++ lib/event_processor/event_dispatcher/send_beacon_dispatcher.browser.spec.ts
