+++ lib/event_processor/event_processor.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { ConversionEvent, ImpressionEvent } from './event_builder/user_event'
-import { LogEvent } from './event_dispatcher'
+import { LogEvent } from './event_dispatcher/event_dispatcher'
 import { getLogger } from '../modules/logging'
 import { Service } from '../service'
 import { Consumer, Fn } from '../utils/type';
