+++ lib/event_processor/event_processor_factory.react_native.spec.ts
@@ -67,7 +67,7 @@ async function mockRequireNetInfo() {
 
 import { createForwardingEventProcessor, createBatchEventProcessor } from './event_processor_factory.react_native';
 import { getForwardingEventProcessor } from './forwarding_event_processor';
-import defaultEventDispatcher from './default_dispatcher.browser';
+import defaultEventDispatcher from './event_dispatcher/default_dispatcher.browser';
 import { EVENT_STORE_PREFIX, FAILED_EVENT_RETRY_INTERVAL } from './event_processor_factory';
 import { getBatchEventProcessor } from './event_processor_factory';
 import { AsyncCache, AsyncPrefixCache, SyncCache, SyncPrefixCache } from '../utils/cache/cache';
