+++ lib/event_processor/event_processor_factory.react_native.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 import { getForwardingEventProcessor } from './forwarding_event_processor';
-import { EventDispatcher } from './event_dispatcher';
+import { EventDispatcher } from './event_dispatcher/event_dispatcher';
 import { EventProcessor } from './event_processor';
-import defaultEventDispatcher from './default_dispatcher.browser';
+import defaultEventDispatcher from './event_dispatcher/default_dispatcher.browser';
 import { BatchEventProcessorOptions, getBatchEventProcessor, getPrefixEventStore } from './event_processor_factory';
 import { EVENT_STORE_PREFIX, FAILED_EVENT_RETRY_INTERVAL } from './event_processor_factory';
 import { AsyncPrefixCache } from '../utils/cache/cache';
