+++ lib/event_processor/forwarding_event_processor.spec.ts
@@ -16,8 +16,8 @@
 import { expect, describe, it, vi } from 'vitest';
 
 import { getForwardingEventProcessor } from './forwarding_event_processor';
-import { EventDispatcher } from './event_dispatcher';
-import { buildLogEvent, makeEventBatch } from './event_builder/build_event_v1';
+import { EventDispatcher } from './event_dispatcher/event_dispatcher';
+import { buildLogEvent, makeEventBatch } from './event_builder/log_event';
 import { createImpressionEvent } from '../tests/mock/create_event';
 import { ServiceState } from '../service';
