+++ lib/index.browser.ts
@@ -18,8 +18,8 @@ import logHelper from './modules/logging/logger';
 import { getLogger, setErrorHandler, getErrorHandler, LogLevel } from './modules/logging';
 import configValidator from './utils/config_validator';
 import defaultErrorHandler from './plugins/error_handler';
-import defaultEventDispatcher from './event_processor/default_dispatcher.browser';
-import sendBeaconEventDispatcher from './event_processor/send_beacon_dispatcher.browser';
+import defaultEventDispatcher from './event_processor/event_dispatcher/default_dispatcher.browser';
+import sendBeaconEventDispatcher from './event_processor/event_dispatcher/send_beacon_dispatcher.browser';
 import * as enums from './utils/enums';
 import * as loggerPlugin from './plugins/logger';
 import { createNotificationCenter } from './core/notification_center';
