+++ lib/index.node.ts
@@ -20,7 +20,7 @@ import * as enums from './utils/enums';
 import * as loggerPlugin from './plugins/logger';
 import configValidator from './utils/config_validator';
 import defaultErrorHandler from './plugins/error_handler';
-import defaultEventDispatcher from './event_processor/default_dispatcher.node';
+import defaultEventDispatcher from './event_processor/event_dispatcher/default_dispatcher.node';
 import { createNotificationCenter } from './core/notification_center';
 import { OptimizelyDecideOption, Client, Config } from './shared_types';
 import { NodeOdpManager } from './plugins/odp_manager/index.node';
